server/requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
idna==2.8
inflection==0.3.1
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
openapi-spec-validator==0.2.6
pathlib==1.0.1
pkg-resources==0.0.0
pylint==2.3.1
python-dotenv==0.10.1
PyYAML==5.1
requests==2.21.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.2
typed-ast==1.3.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
